+++ client/src/components/clubs/CreateMeeting.jsx
@@ -2,8 +2,9 @@ import React from 'react';
 import $ from 'jquery';
 import { Link, Redirect } from 'react-router-dom';
 import AlgoliaPlaces from 'algolia-places-react';
-import MeetingSearchbar from './meeting-searchbar.jsx';
-import MeetingSearchResultsPanel from './meeting-search-results-panel.jsx';
+
+import MeetingSearchbar from './MeetingSearchbar';
+import MeetingSearchResultsPanel from './MeetingSearchResultsPanel';
 
 class CreateMeeting extends React.Component {
   constructor(props) {
@@ -24,7 +25,6 @@ class CreateMeeting extends React.Component {
     this.handleSubmit = this.handleSubmit.bind(this);
     this.getBooks = this.getBooks.bind(this);
     this.loadBook = this.loadBook.bind(this);
-
   }
 
   loadBook(bookImage, bookTitle) {
@@ -114,26 +114,28 @@ class CreateMeeting extends React.Component {
 
   render() {
     return (
-      <div id="create-meeting" className="col-md-9">
-        <div className="container cent">
-          <h3>Create New Meeting</h3>
-          <form onSubmit={this.handleSubmit.bind(this)}>
-            <div className="form-group">
-              <input type="text" className="form-control" id="meeting-name" placeholder="Meeting Time & Date" name="meetingTimestamp" value={this.state.meetingTimestamp} onChange={this.onChange} />
-            </div>
-            <div className="form-group">
-              <input type="text" className="form-control" id="meeting-name" placeholder="Meeting Host" name="meetingHost" value={this.state.meetingHost} onChange={this.onChange} />
-            </div>
-            <div className="form-group">
-              <input className="form-control" id="inputClubDescription" rows="3" placeholder="Add a brief description.  Book to be discussed, who will bring snacks, etc." name="meetingNotes" value={this.state.meetingNotes} onChange={this.onChange} />
-            </div>
-            <MeetingSearchbar searchBooks={this.getBooks}/>
-            <MeetingSearchResultsPanel loadBook={this.loadBook} results={this.state.books}/>
-            <div className="form-group">
-              <AlgoliaPlaces placeholder="Meeting address" onChange={e => this.setLocation(e)} />
-            </div>
-            <input type="submit" className="btn btn-primary centered" value="Submit" />
-          </form>
+      <div className="tab-pane fade" id="nav-create-meeting" role="tabpanel" aria-labelledby="nav-create-meeting-tab">
+        <div id="create-meeting" className="col-md-9">
+          <div className="container cent">
+            <h3>Create New Meeting</h3>
+            <form onSubmit={this.handleSubmit.bind(this)}>
+              <div className="form-group">
+                <input type="text" className="form-control" id="meeting-name" placeholder="Meeting Time & Date" name="meetingTimestamp" value={this.state.meetingTimestamp} onChange={this.onChange} />
+              </div>
+              <div className="form-group">
+                <input type="text" className="form-control" id="meeting-name" placeholder="Meeting Host" name="meetingHost" value={this.state.meetingHost} onChange={this.onChange} />
+              </div>
+              <div className="form-group">
+                <input className="form-control" id="inputClubDescription" rows="3" placeholder="Add a brief description.  Book to be discussed, who will bring snacks, etc." name="meetingNotes" value={this.state.meetingNotes} onChange={this.onChange} />
+              </div>
+              <MeetingSearchbar searchBooks={this.getBooks}/>
+              <MeetingSearchResultsPanel loadBook={this.loadBook} results={this.state.books}/>
+              <div className="form-group">
+                <AlgoliaPlaces placeholder="Meeting address" onChange={e => this.setLocation(e)} />
+              </div>
+              <input type="submit" className="btn btn-primary centered" value="Submit" />
+            </form>
+          </div>
         </div>
       </div>
     );
