+++ client/src/components/clubs/MeetingSearchResultsPanel.jsx
