+++ client/src/components/dashboard/ClubsNearYouPanel.jsx
