+++ client/src/components/dashboard/SearchResultsPanel.jsx
